init and replace
